Do not require cgroups capabilities on windows to run the integration tests.
